fix: a bug where we would drop data during encoding
This one only hit with a specific combination of bytes and was caused due to some optimizing the encoding process. To mitigate this I’m going to add code coverage tools (which may have helped, but in any case are nice to have) and have already added both random and generated byte tests. No longer are we testing just printable ascii.